split Play.tsx into separate components
