+++ client/index.html
@@ -44,7 +44,7 @@
       async
       defer
       data-domain="buzzwords.gg"
-      src="https://analytics.chuckdries.com/js/plausible.local.js"
+      src="https://analytics.chuckdries.com/js/plausible.js"
     ></script>
   </head>
   <body>
