+++ client/package.json
@@ -50,6 +50,7 @@
     "ramda": "^0.27.1",
     "react": "^18.2.0",
     "react-aria": "^3.18.0",
+    "react-cool-dimensions": "^2.0.7",
     "react-div-100vh": "^0.7.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
